tests/wpt/tests/css/css-sizing/contain-intrinsic-size/parsing/from-element-computed.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic Size Test: intrinsic-size with from-element</title>
<link rel="author" href="mailto:kojii@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#intrinsic-size-override">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<script>
test_computed_value("contain-intrinsic-width", "from-element none");
test_computed_value("contain-intrinsic-height", "from-element none");

test_computed_value("contain-intrinsic-inline-size", "from-element none");
test_computed_value("contain-intrinsic-block-size", "from-element none");

test_computed_value("contain-intrinsic-width", "from-element 200px");
test_computed_value("contain-intrinsic-height", "from-element 200px");

test_computed_value("contain-intrinsic-size", "from-element none");
test_computed_value("contain-intrinsic-size", "from-element 200px");
test_computed_value("contain-intrinsic-size", "from-element 300px 200px");
test_computed_value("contain-intrinsic-size", "300px from-element 200px");
</script>
</body>
</html>
